+++ packages/create-pantheon-decoupled-kit/index.ts
@@ -1,0 +1,4 @@
+export * from './src/index';
+export * from './src/types';
+export * from './src/actions';
+export * from './src/utils';