+++ packages/create-pantheon-decoupled-kit/src/generators/index.ts
@@ -1,2 +1,4 @@
 import { nextWp } from './next-wp.generator';
-export const decoupledKitGenerators = [nextWp];
+import { nextDrupal } from './next-drupal.generator';
+import { gatsbyWp } from './gatsby-wp.generator';
+export const decoupledKitGenerators = [nextWp, nextDrupal, gatsbyWp];