+++ packages/create-pantheon-decoupled-kit/src/utils/index.ts
@@ -1,0 +1,2 @@
+export { getPartials } from './getPartials';
+export * from './handlebarsHelpers';